requirements.txt
ruff==0.8.0
Django==5.1.3
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2
psycopg2_binary
drf_spectacular==0.28.0